1440: Allow Consent API controllers to be configured for specific OB versions
https://github.com/SecureApiGateway/SecureApiGateway/issues/1440
