+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/funds/v3_1_10/FundsConfirmationConsentApiController.java
@@ -15,98 +15,24 @@
  */
 package com.forgerock.sapi.gateway.rcs.consent.store.api.funds.v3_1_10;
 
-import java.util.Objects;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.RejectConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.funds.v3_1_10.AuthoriseFundsConfirmationConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.funds.v3_1_10.CreateFundsConfirmationConsentRequest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.funds.v3_1_10.FundsConfirmationConsent;
-import com.forgerock.sapi.gateway.rcs.consent.store.repo.entity.funds.FundsConfirmationConsentEntity;
-import com.forgerock.sapi.gateway.rcs.consent.store.repo.service.funds.FundsConfirmationAuthoriseConsentArgs;
+import com.forgerock.sapi.gateway.rcs.consent.store.api.funds.BaseFundsConfirmationConsentApiController;
 import com.forgerock.sapi.gateway.rcs.consent.store.repo.service.funds.FundsConfirmationConsentService;
 import com.forgerock.sapi.gateway.uk.common.shared.api.meta.obie.OBVersion;
 
-import uk.org.openbanking.datamodel.v3.common.OBExternalRequestStatus1Code;
+import io.swagger.annotations.Api;
 
 @Controller
-public class FundsConfirmationConsentApiController implements FundsConfirmationConsentApi {
-
-    private final Logger logger = LoggerFactory.getLogger(getClass());
-
-    private final FundsConfirmationConsentService consentService;
-
-    private final OBVersion obVersion;
+@Api(tags = {"v3.1.10"})
+@RequestMapping(value = "/consent/store/v3.1.10")
+public class FundsConfirmationConsentApiController extends BaseFundsConfirmationConsentApiController {
 
     @Autowired
     public FundsConfirmationConsentApiController(FundsConfirmationConsentService consentService) {
-        this(consentService, OBVersion.v3_1_10);
+        super(consentService, OBVersion.v3_1_10);
     }
 
-    public FundsConfirmationConsentApiController(FundsConfirmationConsentService consentService, OBVersion obVersion) {
-        this.consentService = Objects.requireNonNull(consentService, "consentService must be provided");
-        this.obVersion = Objects.requireNonNull(obVersion, "obVersion must be provided");
-    }
-
-    @Override
-    public ResponseEntity<FundsConfirmationConsent> createConsent(CreateFundsConfirmationConsentRequest request) {
-        logger.info("Attempting to createConsent: {}", request);
-        final FundsConfirmationConsentEntity consentEntity = new FundsConfirmationConsentEntity();
-        consentEntity.setRequestObj(request.getConsentRequest());
-        consentEntity.setApiClientId(request.getApiClientId());
-        consentEntity.setRequestVersion(obVersion);
-        consentEntity.setStatus(OBExternalRequestStatus1Code.AWAITINGAUTHORISATION.toString());
-
-        final FundsConfirmationConsentEntity persistedEntity = consentService.createConsent(consentEntity);
-        logger.info("Consent created with id: {}", persistedEntity.getId());
-
-        return new ResponseEntity<>(convertEntityToDto(persistedEntity), HttpStatus.CREATED);
-    }
-
-    @Override
-    public ResponseEntity<FundsConfirmationConsent> getConsent(String consentId, String apiClientId) {
-        logger.info("Attempting to getConsent - id: {}, for apiClientId: {}", consentId, apiClientId);
-        return ResponseEntity.ok(convertEntityToDto(consentService.getConsent(consentId, apiClientId)));
-    }
-
-    @Override
-    public ResponseEntity<FundsConfirmationConsent> authoriseConsent(String consentId, AuthoriseFundsConfirmationConsentRequest request) {
-        logger.info("Attempting to authoriseConsent - id: {}, request: {}", consentId, request);
-        final FundsConfirmationAuthoriseConsentArgs authoriseArgs = new FundsConfirmationAuthoriseConsentArgs(consentId, request.getApiClientId(),
-                request.getResourceOwnerId(), request.getAuthorisedDebtorAccountId());
-        return ResponseEntity.ok(convertEntityToDto(consentService.authoriseConsent(authoriseArgs)));
-    }
-
-    @Override
-    public ResponseEntity<FundsConfirmationConsent> rejectConsent(String consentId, RejectConsentRequest request) {
-        logger.info("Attempting to rejectConsent - id: {}, request: {}", consentId, request);
-        return ResponseEntity.ok(convertEntityToDto(consentService.rejectConsent(consentId, request.getApiClientId(), request.getResourceOwnerId())));
-    }
-
-    @Override
-    public ResponseEntity<Void> deleteConsent(String consentId, String apiClientId) {
-        logger.info("Attempting to deleteConsent - id: {}, apiClientId: {}", consentId, apiClientId);
-        consentService.deleteConsent(consentId, apiClientId);
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    private FundsConfirmationConsent convertEntityToDto(FundsConfirmationConsentEntity entity) {
-        final FundsConfirmationConsent dto = new FundsConfirmationConsent();
-        dto.setApiClientId(entity.getApiClientId());
-        dto.setId(entity.getId());
-        dto.setCreationDateTime(entity.getCreationDateTime());
-        dto.setStatus(entity.getStatus());
-        dto.setStatusUpdateDateTime(entity.getStatusUpdatedDateTime());
-        dto.setRequestVersion(entity.getRequestVersion());
-        dto.setRequestObj(entity.getRequestObj());
-        dto.setResourceOwnerId(entity.getResourceOwnerId());
-        dto.setAuthorisedDebtorAccountId(entity.getAuthorisedDebtorAccountId());
-        return dto;
-    }
 }
