+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/payment/vrp/v3_1_10/DomesticVRPConsentApiController.java
@@ -22,6 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.forgerock.sapi.gateway.rcs.consent.store.api.payment.vrp.BaseDomesticVRPConsentApiController;
 import com.forgerock.sapi.gateway.rcs.consent.store.repo.service.payment.vrp.DomesticVRPConsentService;
 import com.forgerock.sapi.gateway.uk.common.shared.api.meta.obie.OBVersion;
 
@@ -33,7 +34,7 @@ import io.swagger.annotations.Api;
 @Controller
 @Api(tags = {"v3.1.10"})
 @RequestMapping(value = "/consent/store/v3.1.10")
-public class DomesticVRPConsentApiController extends com.forgerock.sapi.gateway.rcs.consent.store.api.payment.vrp.BaseDomesticVRPConsentApiController {
+public class DomesticVRPConsentApiController extends BaseDomesticVRPConsentApiController {
 
     @Autowired
     public DomesticVRPConsentApiController(DomesticVRPConsentService consentService,
